Currency System Works
Now, the player has money. The money is displayed as a UI graphic in the bottom right corner. All items can be toggled to be free or a shop item. The player is notified above their coin UI if they cannot afford an item, and how much it costs. There are three coins, gold, silver and bronze. Bronze- 1, Silver- 5, and Gold-10.
I have done some minimal testing, so there still may be bugs.